backend/requirements.txt
django
django-cors-headers
django-ninja
djangorestframework 
djangorestframework-simplejwt
psycopg2-binary
pytest
pytest-django